Just made reset chain smarter. It now checks if the chain is already empty before trying to reset it. If it is empty, it will not try to reset it. This is to prevent the error message "Error: No such container: chain" from showing up.

+++ resetChain.sh
@@ -1,5 +1,12 @@
 #!/bin/zsh
-rm -f ./blk/*.blk
-rm -f chain
+if test -f ./chain; then
+  rm -f chain
+fi
+
+if test -z "$(ls -A ./blk)"; then
+else
+  rm -f ./blk/*.blk
+fi
+
 echo "Genesis Block the unmoved mover" > chain
